+++ src/app/src/main/kotlin/org/imperial/mrc/hint/security/oauth2/OAuth2LogoutActionBuilder.kt
@@ -1,0 +1,26 @@
+package org.imperial.mrc.hint.security.oauth2
+
+import org.imperial.mrc.hint.AppProperties
+import org.pac4j.core.context.WebContext
+import org.pac4j.core.context.session.SessionStore
+import org.pac4j.core.exception.http.RedirectionAction
+import org.pac4j.core.logout.LogoutActionBuilder
+import org.pac4j.core.profile.UserProfile
+import org.pac4j.core.util.HttpActionHelper
+import java.util.*
+
+class OAuth2LogoutActionBuilder(
+    appProperties: AppProperties,
+    oauth2State: OAuth2State,
+) : LogoutActionBuilder, OAuth2AuthenticationRedirection(appProperties, oauth2State)
+{
+    override fun getLogoutAction(
+        context: WebContext?,
+        sessionStore: SessionStore?,
+        currentProfile: UserProfile?,
+        targetUrl: String?,
+    ): Optional<RedirectionAction>
+    {
+        return Optional.of(HttpActionHelper.buildRedirectUrlAction(context, oauth2LogoutRedirect()))
+    }
+}