+++ src/app/src/test/kotlin/org/imperial/mrc/hint/helpers/AuthInterceptor.kt
@@ -23,7 +23,7 @@ class AuthInterceptor(restTemplate: TestRestTemplate, password: String = "passwo
         val headers = HttpHeaders()
         headers.contentType = MediaType.APPLICATION_FORM_URLENCODED
 
-        val entity = restTemplate.postForEntity<String>("/callback/", HttpEntity(map, headers))
+        val entity = restTemplate.postForEntity<String>("/callback/formClient", HttpEntity(map, headers))
 
         assertThat(entity.statusCode).isEqualTo(HttpStatus.SEE_OTHER)
         assertThat(entity.headers.location!!.toString()).isEqualTo("/")