+++ NEWS.md
@@ -1,6 +1,15 @@
-# hint 2.13.0
+# hint 2.15.0
 
 * Log elapsed time for ADR responses
+
+# hint 2.14.0
+
+* Add Auth0 login method
+
+# hint 2.13.0
+
+* Update hint to work with options schema changes
+>>>>>>> master
 
 # hint 2.12.0